extract slideout into component
