+++ src/app/component/index.js
@@ -14,6 +14,7 @@ import UserPhoneUnverifiedPopup from "./UserPhoneUnverifiedPopup";
 import ContactComponent from "./ContactComponent";
 import { Body1, Body2, H1, H2, H3, H4, H5, H6, Div, TypographyWrapper } from "./Typography";
 import ComingSoon from "./ComingSoon/ComingSoon";
+import SlideUpContainer from "./SlideUpContainer";
 
 export {
   AddressInput,
@@ -26,6 +27,7 @@ export {
   MissionList,
   DownloadMissionsCsv,
   ShowDeliveryRoute,
+  SlideUpContainer,
   Popup,
   UserPhoneUnverifiedPopup,
   ContactComponent,
